requirements/requirements-m1pro.txt
# jax must be installed previously depending on hardware
# jax==0.4.6 is the oldest version available now (2024.11.03)
jax<0.5
dm-haiku
flax
typing_extensions>=4.2.0
svgwrite
pytest